backend/services/user-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
mangum
pydantic
python-multipart
passlib[bcrypt]
python-jose
